[WIP] memchr memmove strtrim putendl_fd

+++ ft_calloc.c
@@ -6,12 +6,12 @@
 /*   By: juggorr <juggorr@gmail.com>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 17:50:28 by juggorr           #+#    #+#             */
-/*   Updated: 2023/12/05 18:11:43 by juggorr          ###   ########.fr       */
+/*   Updated: 2023/12/06 19:23:39 by junghopa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
-extern void	ft_bzero(void *b, size_t len);
+extern void	ft_bzero(void *b, unsigned int len);
 
 void	*ft_calloc(size_t number, size_t size)
 {
