+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: juggorr <juggorr@gmail.com>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/30 15:17:40 by juggorr           #+#    #+#             */
-/*   Updated: 2023/12/05 19:24:54 by juggorr          ###   ########.fr       */
+/*   Updated: 2023/12/06 19:23:01 by junghopa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef LIBFT_H
@@ -16,7 +16,7 @@
 # include <stdlib.h>
 
 int				ft_atoi(char const *s1);
-void			ft_bzero(void *s, unsigned long n);
+void			ft_bzero(void *s, unsigned int n);
 void			*ft_calloc(size_t number, size_t size);
 int				ft_isalnum(int c);
 int				ft_isalpha(int c);
@@ -25,10 +25,10 @@ int				ft_isdigit(int c);
 int				ft_isprint(int c);
 char			*ft_itoa(int n);
 char			ft_digits_cnt(int n);
-void			ft_memchr(const void *s, int c, unsigned long n);
-int				ft_memcmp(const void *s1, const void *s2, unsigned long n);
-void			*ft_memcpy(void *dst, void *src, unsigned long n);
-void			*ft_memset(void *b, int c, unsigned long len);
+void			ft_memchr(const void *s, int c, unsigned int n);
+int				ft_memcmp(const void *s1, const void *s2, unsigned int n);
+void			*ft_memcpy(void *dst, void *src, unsigned int n);
+void			*ft_memset(void *b, int c, unsigned int len);
 void			*ft_memmove(void *dst, void const *src, unsigned int len);
 void			ft_putchar_fd(char c, int fd);
 void			ft_putendl_fd(char *s, int fd);
@@ -43,8 +43,8 @@ unsigned long	ft_strlcat(char *dset, char *src, unsigned long size);
 unsigned int	ft_strlcpy(char *dst, char *str, unsigned int size);
 unsigned int	ft_strlen(const char *s);
 char			*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-int				ft_strncmp(const char *s1, const char *s2, unsigned long n);
-char			*ft_strnstr(const char *s1, const char *s2, unsigned long n);
+int				ft_strncmp(const char *s1, const char *s2, unsigned int n);
+char			*ft_strnstr(const char *s1, const char *s2, unsigned int n);
 char			*ft_strrchr(char *s, int c);
 char			*ft_strtrim(char const *s1, char const *set);
 char			*ft_substr(char const *s, unsigned int start, size_t len);
